Div generation & content done
Prints semesters and classes to a div that it tosses in at the top. CSS
styling for this div is also done now.

The following still needs to be done: actually insert the stylesheet for
the new stuff, optionally, append even and odd classes to the divs for
each class so that they alternate color, grab the person's name and add
it to the top